Add files via upload
If I follow the logic in the instruction set and take the first occurrence of each employee and then filter out the inactive employees, then I shouldn't be getting the employee with ID = 12155. However, the only way I found to get around this is by selecting the "from_date" and "to_date" columns from the titles table instead of the dept_emp table. I also created intermediary tables m1 and m2 to make this logic work, hence why my sql file contains those queries. Screenshots for the 12155 case have been attached to the repo in addition to my sql file and the exported tables.