Add files via upload
If I follow the logic and take the first occurrence of each employee and then filter out the inactive employees, then I shouldn't be getting 12155. However, the only way I found to get around this is by selecting the "from_date" and "to_date" columns from the titles table instead of the dept_emp table. The attached screenshot show the 12155 case. 
Please note that as a result of this my sql file had to be slightly modified and intermediary tables m1 and m2 were created in order to get the correct result.